Mejora la creacion de Link de vi -> vim

+++ INSTALL.sh
@@ -16,7 +16,11 @@ function INSTALAR {
      echo ""
      sudo mkdir /usr/share/fonts/truetype/Ubuntu__Mono_Nerd_Font
      sudo cp Tema/ttf/*.ttf /usr/share/fonts/truetype/Ubuntu__Mono_Nerd_Font/
-     sudo ln -s /bin/vim /bin/vi
+     
+     if ( -e /usr/vi ) ; then ; echo "" ; else
+         sudo ln -s /bin/vim /bin/vi
+     fi
+
      echo ""
      echo "           Se ha instalado Correctamente en el VIM del Usuario. Terminando el Instalador..."
      echo "           Solamente queda una cosa..."
